improved overall structure

+++ src/FEM/ParamDataStructures/ParamSpace.jl
@@ -16,7 +16,7 @@ _get_params(r::ParamRealization) = r.params # this function should stay local
 _get_params(r::TrivialParamRealization) = [r.params] # this function should stay local
 num_params(r::ParamRealization) = length(_get_params(r))
 Base.length(r::ParamRealization) = num_params(r)
-Base.getindex(r::ParamRealization,i::Integer) = ParamRealization(getindex(_get_params(r),i))
+Base.getindex(r::ParamRealization,i) = ParamRealization(getindex(_get_params(r),i))
 Base.copy(r::ParamRealization) = ParamRealization(copy(_get_params(r)))
 Arrays.testitem(r::ParamRealization) = testitem(_get_params(r))
